Creates two flavors for easier publication of BillingClient 3 compatible library (#444)
* creates two flavors

* extracts 3.0.2 as a variable

* fix compilation

* fix unresolved skus and sku

* adds deploys

* change to compileOnly so we prevent the duplicated classes errors

* reorder and fix skus size check

* moves flavors to base-application

* updates flavor versions

* fix names of packages

* changes names and adds Amazon

* fixing deploys

* fix some errors

* removes unused mocks

* fixes purchasestests

* Run PostingTransactionsTests only in latest dependencies

* moves GoogleDeviceIdentifiersFetcherTests back

* renames files to reflect billingclient version

* more renames

* adds missing import
