+++ build.gradle
@@ -4,6 +4,7 @@ buildscript {
     ext.buildToolsVersion = "30.0.3"
     ext.minVersion = 14
     ext.billingVersion = "4.0.0"
+    ext.billingClient3Version = "3.0.2"
     ext.lifecycleVersion = "2.3.0-rc01"
     repositories {
         mavenCentral()
