+++ common/src/main/java/com/revenuecat/purchases/common/errors.kt
@@ -1,6 +1,5 @@
 package com.revenuecat.purchases.common
 
-import com.android.billingclient.api.BillingClient
 import com.revenuecat.purchases.PurchasesError
 import com.revenuecat.purchases.PurchasesErrorCode
 import com.revenuecat.purchases.common.networking.HTTPResult
@@ -89,33 +88,6 @@ private fun BackendErrorCode.toPurchasesErrorCode(): PurchasesErrorCode {
         BackendErrorCode.BackendInternalServerError -> PurchasesErrorCode.UnexpectedBackendResponseError
         BackendErrorCode.BackendProductIDsMalformed -> PurchasesErrorCode.UnsupportedError
     }
-}
-
-internal fun @receiver:BillingClient.BillingResponseCode Int.getBillingResponseCodeName(): String {
-    val allPossibleBillingResponseCodes = BillingClient.BillingResponseCode::class.java.declaredFields
-    return allPossibleBillingResponseCodes
-        .firstOrNull { it.getInt(it) == this }
-        ?.name
-        ?: "$this"
-}
-
-fun Int.billingResponseToPurchasesError(underlyingErrorMessage: String): PurchasesError {
-    val errorCode = when (this) {
-        BillingClient.BillingResponseCode.BILLING_UNAVAILABLE,
-        BillingClient.BillingResponseCode.ITEM_NOT_OWNED,
-        BillingClient.BillingResponseCode.FEATURE_NOT_SUPPORTED -> PurchasesErrorCode.PurchaseNotAllowedError
-        BillingClient.BillingResponseCode.ERROR,
-        BillingClient.BillingResponseCode.SERVICE_UNAVAILABLE,
-        BillingClient.BillingResponseCode.SERVICE_DISCONNECTED,
-        BillingClient.BillingResponseCode.SERVICE_TIMEOUT -> PurchasesErrorCode.StoreProblemError
-        BillingClient.BillingResponseCode.OK -> PurchasesErrorCode.UnknownError
-        BillingClient.BillingResponseCode.USER_CANCELED -> PurchasesErrorCode.PurchaseCancelledError
-        BillingClient.BillingResponseCode.ITEM_UNAVAILABLE -> PurchasesErrorCode.ProductNotAvailableForPurchaseError
-        BillingClient.BillingResponseCode.DEVELOPER_ERROR -> PurchasesErrorCode.PurchaseInvalidError
-        BillingClient.BillingResponseCode.ITEM_ALREADY_OWNED -> PurchasesErrorCode.ProductAlreadyPurchasedError
-        else -> PurchasesErrorCode.UnknownError
-    }
-    return PurchasesError(errorCode, underlyingErrorMessage)
 }
 
 data class SubscriberAttributeError(
