+++ fastlane/Fastfile
@@ -108,13 +108,27 @@ platform :android do
       "mavenCentralPassword" => ENV['SONATYPE_NEXUS_TOKEN_PASSWORD'],
       "RELEASE_SIGNING_ENABLED" => true
     }
-    puts "Deploying #{version}"
+    UI.verbose("Deploying #{version}")
     gradle(
       tasks: [
         "androidSourcesJar", "androidJavadocJar", "publish --no-daemon --no-parallel"
       ],
       properties: gradleProperties
     )
+
+    UI.verbose("Creating special version for Unity IAP with BillingClient 3 and Amazon 2 for version: #{version}")
+    gradleProperties["PUBLISH_VARIANT"] = "unityIAPRelease"
+    
+    UI.verbose("Adding -unityiap to artifact ids")
+    
+
+    gradle(
+      tasks: [
+        "androidSourcesJar", "androidJavadocJar", "publish --no-daemon --no-parallel"
+      ],
+      properties: gradleProperties
+    )
+    
     unless is_snapshot_version?(version)
       gradle(
         tasks: [
