+++ feature/amazon/build.gradle
@@ -2,6 +2,8 @@ apply from: "$rootProject.projectDir/library.gradle"
 
 ext {
     amazonVersion = "3.0.2"
+    amazon2Version = "2.0.76"
+    amazon2FileName = "in-app-purchasing-${amazon2Version}.jar"
 }
 
 dependencies {
@@ -12,7 +14,8 @@ dependencies {
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlinVersion"
 
-    implementation "com.amazon.device:amazon-appstore-sdk:$amazonVersion"
+    latestDependenciesImplementation "com.amazon.device:amazon-appstore-sdk:$amazonVersion"
+    unityIAPCompileOnly files("libs/in-app-purchasing-${amazon2Version}.jar")
 
     testImplementation "com.amazon.device:amazon-appstore-sdk:$amazonVersion"
     testImplementation project(":test-utils")
@@ -21,3 +24,46 @@ dependencies {
 // Not running with this flag causes verification issues due to the way the Amazon jar is compiled
 // https://github.com/robolectric/robolectric-gradle-plugin/issues/144#issuecomment-265899560
 tasks.withType(Test) { jvmArgs += "-noverify" }
+
+task getAmazonLibrary {
+    ext {
+        downloadURL = "https://amzndevresources.com/iap/sdk/AmazonInAppPurchasing_Android.zip"
+        destFile = new File(projectDir, "libs/$amazon2FileName")
+    }
+
+    inputs.property('downloadURL', downloadURL)
+    inputs.property('fileToExtract', amazon2FileName)
+    outputs.file(destFile)
+
+    doLast {
+        if (!destFile.exists()) {
+            println 'Downloading Amazon dependency'
+            File destDir = destFile.parentFile
+            destDir.mkdirs()
+
+            File downloadFile = new File(temporaryDir, 'download.zip')
+            new URL(downloadURL).withInputStream { is ->
+                downloadFile.withOutputStream { it << is }
+            }
+
+            project.copy {
+                from {
+                    zipTree(downloadFile).matching { include "**/$amazon2FileName" }.singleFile
+                }
+
+                into(destDir)
+            }
+        }
+    }
+}
+
+task cleanAmazonLibrary(type: Delete) {
+    delete new File(projectDir, "libs/$amazon2FileName")
+}
+
+afterEvaluate {
+    preUnityIAPDebugBuild.dependsOn getAmazonLibrary
+    preUnityIAPReleaseBuild.dependsOn getAmazonLibrary
+    preLatestDependenciesDebugBuild.dependsOn cleanAmazonLibrary
+    preLatestDependenciesReleaseBuild.dependsOn cleanAmazonLibrary
+}
