+++ feature/google/build.gradle
@@ -11,7 +11,8 @@ dependencies {
     implementation 'com.google.android.gms:play-services-ads-identifier:17.0.1'
     implementation "androidx.core:core-ktx:1.3.1"
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlinVersion"
-    implementation "com.android.billingclient:billing:$billingVersion"
+    latestDependenciesImplementation "com.android.billingclient:billing:$billingVersion"
+    unityIAPCompileOnly "com.android.billingclient:billing:$billingClient3Version"
 
     testImplementation project(":test-utils")
     testImplementation 'com.google.android.gms:play-services-ads-identifier:17.0.1'
