+++ feature/google/src/main/java/com/revenuecat/purchases/google/billingResultExtensions.kt
@@ -1,0 +1,9 @@
+package com.revenuecat.purchases.google
+
+import com.android.billingclient.api.BillingClient
+import com.android.billingclient.api.BillingResult
+
+fun BillingResult.toHumanReadableDescription() =
+    "DebugMessage: $debugMessage. ErrorCode: ${responseCode.getBillingResponseCodeName()}."
+
+fun BillingResult.isSuccessful() = responseCode == BillingClient.BillingResponseCode.OK
