+++ feature/google/src/main/java/com/revenuecat/purchases/google/paymenTransactionConversions.kt
@@ -3,6 +3,7 @@ package com.revenuecat.purchases.google
 import com.android.billingclient.api.Purchase
 import com.android.billingclient.api.PurchaseHistoryRecord
 import com.revenuecat.purchases.ProductType
+import com.revenuecat.purchases.common.listOfSkus
 import com.revenuecat.purchases.models.StoreTransaction
 import com.revenuecat.purchases.models.PurchaseType
 import com.revenuecat.purchases.models.PurchaseState
@@ -13,7 +14,7 @@ fun Purchase.toStoreTransaction(
     presentedOfferingIdentifier: String?
 ): StoreTransaction = StoreTransaction(
     orderId = this.orderId,
-    skus = this.skus,
+    skus = this.listOfSkus,
     type = productType,
     purchaseTime = this.purchaseTime,
     purchaseToken = this.purchaseToken,
@@ -37,7 +38,7 @@ fun PurchaseHistoryRecord.toStoreTransaction(
 ): StoreTransaction {
     return StoreTransaction(
         orderId = null,
-        skus = this.skus,
+        skus = this.listOfSkus,
         type = type,
         purchaseTime = this.purchaseTime,
         purchaseToken = this.purchaseToken,
