+++ gradle.properties
@@ -21,3 +21,4 @@ android.enableJetifier=true
 
 #Do not sign releases. When calling uploadArchives pass -PRELEASE_SIGNING_ENABLED=true
 RELEASE_SIGNING_ENABLED=false
+PUBLISH_VARIANT=latestDependenciesRelease
