+++ purchases/src/testLatestDependencies/java/com/revenuecat/purchases/PostingTransactionsTestsBillingClient4.kt
@@ -301,5 +301,4 @@ class PostingTransactionsTests {
             )
         }
     }
-
 }
